# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface SendOpenaiVoiceMessageBody {
  /** Base64-encoded audio data */
  audio: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export interface TranscribeAudioBody {
  /** Base64-encoded audio */
  audio: string;
}

export interface TranscribeAudioResponse {
  text: string;
}

export interface TextToSpeechBody {
  text: string;
  voice?: string;
}

export interface TextToSpeechResponse {
  /** Base64-encoded audio */
  audio: string;
}

export interface StudyMaterial {
  id: number;
  title: string;
  subject: string;
  content: string;
  summary?: string;
  fileType: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateStudyMaterialBody {
  title: string;
  subject: string;
  content: string;
  fileType: string;
}

export type GenerateQuizBodyDifficulty =
  (typeof GenerateQuizBodyDifficulty)[keyof typeof GenerateQuizBodyDifficulty];

export const GenerateQuizBodyDifficulty = {
  makkelijk: "makkelijk",
  gemiddeld: "gemiddeld",
  moeilijk: "moeilijk",
} as const;

export interface GenerateQuizBody {
  difficulty: GenerateQuizBodyDifficulty;
  numQuestions?: number;
}

export interface QuizQuestion {
  question: string;
  options?: string[];
  correctAnswer: string;
  explanation?: string;
}

export interface Quiz {
  id: number;
  materialId: number;
  questions: QuizQuestion[];
  createdAt: string;
}

export interface Exercise {
  id: number;
  materialId: number;
  question: string;
  answer: string;
  difficulty: string;
  tip?: string;
}

export type StudyGoalStatus =
  (typeof StudyGoalStatus)[keyof typeof StudyGoalStatus];

export const StudyGoalStatus = {
  actief: "actief",
  voltooid: "voltooid",
  gepauzeerd: "gepauzeerd",
} as const;

export interface StudyGoal {
  id: number;
  title: string;
  subject: string;
  targetDate: string;
  hoursPerWeek: number;
  progress: number;
  status: StudyGoalStatus;
  createdAt: string;
}

export interface CreateStudyGoalBody {
  title: string;
  subject: string;
  targetDate: string;
  hoursPerWeek: number;
}

export type UpdateStudyGoalBodyStatus =
  (typeof UpdateStudyGoalBodyStatus)[keyof typeof UpdateStudyGoalBodyStatus];

export const UpdateStudyGoalBodyStatus = {
  actief: "actief",
  voltooid: "voltooid",
  gepauzeerd: "gepauzeerd",
} as const;

export interface UpdateStudyGoalBody {
  title?: string;
  subject?: string;
  targetDate?: string;
  hoursPerWeek?: number;
  progress?: number;
  status?: UpdateStudyGoalBodyStatus;
}

export type CalendarEventType =
  (typeof CalendarEventType)[keyof typeof CalendarEventType];

export const CalendarEventType = {
  studie: "studie",
  afspraak: "afspraak",
  vrij: "vrij",
} as const;

export interface CalendarEvent {
  id: number;
  title: string;
  description?: string;
  startTime: string;
  endTime: string;
  type: CalendarEventType;
  subject?: string;
  color?: string;
  completed: boolean;
}

export type CreateCalendarEventBodyType =
  (typeof CreateCalendarEventBodyType)[keyof typeof CreateCalendarEventBodyType];

export const CreateCalendarEventBodyType = {
  studie: "studie",
  afspraak: "afspraak",
  vrij: "vrij",
} as const;

export interface CreateCalendarEventBody {
  title: string;
  description?: string;
  startTime: string;
  endTime: string;
  type: CreateCalendarEventBodyType;
  subject?: string;
  color?: string;
}

export interface UpdateCalendarEventBody {
  title?: string;
  description?: string;
  startTime?: string;
  endTime?: string;
  completed?: boolean;
}

export interface RescheduleWeekBody {
  weekStartDate: string;
}

export interface SubjectProgress {
  subject: string;
  hoursStudied: number;
  progressPercent: number;
  score: number;
}

export interface ProgressSummary {
  totalStudyHours: number;
  weeklyHours: number;
  completedGoals: number;
  totalGoals: number;
  completedTasks: number;
  totalTasks: number;
  averageScore: number;
  studiedSubjects: SubjectProgress[];
}

export type WeakPointSeverity =
  (typeof WeakPointSeverity)[keyof typeof WeakPointSeverity];

export const WeakPointSeverity = {
  laag: "laag",
  gemiddeld: "gemiddeld",
  hoog: "hoog",
} as const;

export interface WeakPoint {
  id: number;
  subject: string;
  topic: string;
  description: string;
  severity: WeakPointSeverity;
  suggestedAction: string;
}

export interface WeeklyReviewBody {
  /** User's own reflection text or voice-transcribed text */
  feedback: string;
}

export interface StudyStreak {
  currentStreak: number;
  longestStreak: number;
  lastStudiedDate?: string;
  studiedToday: boolean;
}

export interface NotificationSettings {
  id: number;
  enabled: boolean;
  eveningReminder: boolean;
  eveningReminderTime: string;
  pushSubscription?: string;
}

export interface UpdateNotificationSettingsBody {
  enabled?: boolean;
  eveningReminder?: boolean;
  eveningReminderTime?: string;
  pushSubscription?: string;
}

export type UserSettingsDifficultyLevel =
  (typeof UserSettingsDifficultyLevel)[keyof typeof UserSettingsDifficultyLevel];

export const UserSettingsDifficultyLevel = {
  makkelijk: "makkelijk",
  gemiddeld: "gemiddeld",
  moeilijk: "moeilijk",
} as const;

export type UserSettingsCoachStyle =
  (typeof UserSettingsCoachStyle)[keyof typeof UserSettingsCoachStyle];

export const UserSettingsCoachStyle = {
  streng: "streng",
  gebalanceerd: "gebalanceerd",
  aanmoedigend: "aanmoedigend",
} as const;

export interface UserSettings {
  id: number;
  userName: string;
  difficultyLevel: UserSettingsDifficultyLevel;
  preferredLanguage: string;
  voiceEnabled: boolean;
  coachStyle: UserSettingsCoachStyle;
  studyReminders: boolean;
  weeklyGoalHours: number;
}

export type UpdateUserSettingsBodyDifficultyLevel =
  (typeof UpdateUserSettingsBodyDifficultyLevel)[keyof typeof UpdateUserSettingsBodyDifficultyLevel];

export const UpdateUserSettingsBodyDifficultyLevel = {
  makkelijk: "makkelijk",
  gemiddeld: "gemiddeld",
  moeilijk: "moeilijk",
} as const;

export type UpdateUserSettingsBodyCoachStyle =
  (typeof UpdateUserSettingsBodyCoachStyle)[keyof typeof UpdateUserSettingsBodyCoachStyle];

export const UpdateUserSettingsBodyCoachStyle = {
  streng: "streng",
  gebalanceerd: "gebalanceerd",
  aanmoedigend: "aanmoedigend",
} as const;

export interface UpdateUserSettingsBody {
  userName?: string;
  difficultyLevel?: UpdateUserSettingsBodyDifficultyLevel;
  preferredLanguage?: string;
  voiceEnabled?: boolean;
  coachStyle?: UpdateUserSettingsBodyCoachStyle;
  studyReminders?: boolean;
  weeklyGoalHours?: number;
}

export type ListCalendarEventsParams = {
  startDate?: string;
  endDate?: string;
};