# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StudyFlow AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
  /openai/conversations/{id}/voice-messages:
    post:
      operationId: sendOpenaiVoiceMessage
      tags: [openai]
      summary: Send audio and receive a streaming voice response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiVoiceMessageBody"
      responses:
        "200":
          description: SSE stream of transcript and audio chunks
          content:
            text/event-stream: {}
  /openai/transcribe:
    post:
      operationId: transcribeAudio
      tags: [openai]
      summary: Transcribe audio to text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscribeAudioBody"
      responses:
        "200":
          description: Transcribed text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscribeAudioResponse"
  /openai/tts:
    post:
      operationId: textToSpeech
      tags: [openai]
      summary: Convert text to speech audio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TextToSpeechBody"
      responses:
        "200":
          description: Base64 audio data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TextToSpeechResponse"
...
  /study-goals:
    get:
      operationId: listStudyGoals
      tags: [studygoals]
      summary: List all study goals
      responses:
        "200":
          description: List of study goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyGoal"
    post:
      operationId: createStudyGoal
      tags: [studygoals]
      summary: Create a new study goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudyGoalBody"
      responses:
        "201":
          description: Created study goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyGoal"
  /study-goals/{id}:
    put:
      operationId: updateStudyGoal
      tags: [studygoals]
      summary: Update a study goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudyGoalBody"
      responses:
        "200":
          description: Updated study goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyGoal"
    delete:
      operationId: deleteStudyGoal
      tags: [studygoals]
      summary: Delete a study goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /calendar-events:
    get:
      operationId: listCalendarEvents
      tags: [calendarevents]
      summary: List calendar events
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of calendar events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
    post:
      operationId: createCalendarEvent
      tags: [calendarevents]
      summary: Create a calendar event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCalendarEventBody"
      responses:
        "201":
          description: Created calendar event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
  /calendar-events/{id}:
    put:
      operationId: updateCalendarEvent
      tags: [calendarevents]
      summary: Update a calendar event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCalendarEventBody"
      responses:
        "200":
          description: Updated calendar event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
    delete:
      operationId: deleteCalendarEvent
      tags: [calendarevents]
      summary: Delete a calendar event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /calendar-events/reschedule-week:
    post:
      operationId: rescheduleWeek
      tags: [calendarevents]
      summary: AI reschedules week study blocks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleWeekBody"
      responses:
        "200":
          description: SSE stream of rescheduling updates
          content:
            text/event-stream: {}
...
components:
  schemas:
...
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    SendOpenaiVoiceMessageBody:
      type: object
      properties:
        audio:
          type: string
          description: Base64-encoded audio data
      required:
        - audio
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    TranscribeAudioBody:
      type: object
      properties:
        audio:
          type: string
          description: Base64-encoded audio
      required:
        - audio
    TranscribeAudioResponse:
      type: object
      properties:
        text:
          type: string
      required:
        - text
    TextToSpeechBody:
      type: object
      properties:
        text:
          type: string
        voice:
          type: string
      required:
        - text
    TextToSpeechResponse:
      type: object
      properties:
        audio:
          type: string
          description: Base64-encoded audio
      required:
        - audio
...
    StudyGoal:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subject:
          type: string
        targetDate:
          type: string
          format: date-time
        hoursPerWeek:
          type: number
        progress:
          type: number
        status:
          type: string
          enum: [actief, voltooid, gepauzeerd]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - subject
        - targetDate
        - hoursPerWeek
        - progress
        - status
        - createdAt
    CreateStudyGoalBody:
      type: object
      properties:
        title:
          type: string
        subject:
          type: string
        targetDate:
          type: string
          format: date-time
        hoursPerWeek:
          type: number
      required:
        - title
        - subject
        - targetDate
        - hoursPerWeek
    UpdateStudyGoalBody:
      type: object
      properties:
        title:
          type: string
        subject:
          type: string
        targetDate:
          type: string
          format: date-time
        hoursPerWeek:
          type: number
        progress:
          type: number
        status:
          type: string
          enum: [actief, voltooid, gepauzeerd]
    CalendarEvent:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        type:
          type: string
          enum: [studie, afspraak, vrij]
        subject:
          type: string
        color:
          type: string
        completed:
          type: boolean
      required:
        - id
        - title
        - startTime
        - endTime
        - type
        - completed
    CreateCalendarEventBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        type:
          type: string
          enum: [studie, afspraak, vrij]
        subject:
          type: string
        color:
          type: string
      required:
        - title
        - startTime
        - endTime
        - type
    UpdateCalendarEventBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        completed:
          type: boolean
    RescheduleWeekBody:
      type: object
      properties:
        weekStartDate:
          type: string
      required:
        - weekStartDate
...