# lib/api-zod/src/generated/types/createCalendarEventBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCalendarEventBodyType } from "./createCalendarEventBodyType";

export interface CreateCalendarEventBody {
  title: string;
  description?: string;
  startTime: Date;
  endTime: Date;
  type: CreateCalendarEventBodyType;
  subject?: string;
  color?: string;
}