# lib/api-zod/src/generated/types/rescheduleWeekBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RescheduleWeekBody {
  weekStartDate: string;
}